requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
idna==2.10
oauthlib==3.1.0
py-trello==0.17.1
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
